Split capture normalization modules
